perf(vetkd): Use multi-threading in VetKD payload builder/validator (#7373)
I recommend reviewing this after setting the tick to "Hide whitespace".

## Background
VetKD shares are created and gossiped by the IDKG client in a separate
thread. Once enough valid shares exist, a VetKD response will be created
and included into the block by the blockmaker. This is similar to the
online phases of tECDSA and tSchnorr, however, ECDSA and Schnorr shares
are combined as part of the IDkg payload instead.

Originally, there didn't seem to be a reason to add multi-threading to
the online phases of the chain key protocols (which includes VetKD),
since combining and verifying combined ECDSA and Schnorr signature
shares is quite cheap. Consequently, handling many of them on the
critical path of block making does not significantly affect the
finalization rate.

## Problem
It turns out that combining and verifying VetKD shares is significantly
more expensive compared to tECDSA/tSchnorr. Under high VetKD load, a
drop in finalization rate can be observed.

## Solution
Similar to previous work, this PR adds multi-threading to the VetKD
payload builder and validator. This is done by moving the creation of
the consensus thread pool out of the consensus module, and into the
replica setup. This way, the same thread pool can be passed to both
consensus (where it is used for IDKG), and the VetKD payload builder.

## Performance Impact
Before this PR, the throughput was limited to ~18 VetKD agreements per
second on a 25 node subnet, with the finalization rate dropping to
~0.6/s. Afterwards, with 34 nodes, generating 30 agreements per second
seems to be possible without affecting the finalization rate
significantly (I did not try going higher). Production subnets are
currently still capped to a lower throughput due to smaller configured
queue size.

## Future Work
VetKD shares are currently handled in a different payload section than
ECDSA/Schnorr shares. For simplicity, we plan to move the handling of
ECDSA/Schnorr shares into the same payload as VetKD. At that point,
these protocols will also benefit by the multi-threading added in this
PR.

After this PR, the thread pool is created during the replica setup. This
opens the door for enabling multi-threading also in other payload
builders, i.e. HTTP outcalls.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>