feat(recovery): [CON-1590] download consensus pool and state from different nodes in app subnet recoveries (#7147)
Currently, if the DFINITY-owned node of an app subnet does not have the
latest state, we would either:
- Select an up-to-date node in the DownloadState step, but this is
expensive.
- Manually rsync an up-to-date node’s consensus pool and replay it on
top of the DFINITY-owned node’s state.

The `DownloadConsensusPool` step added as part of NNS recovery exactly
does this second option as part of the tool, so this PR adds it to app
subnet recoveries.

The method for selecting the download and upload nodes in the system
tests has been updated to further closely simulate a real-world
scenario:
- The `download_state_node` is selected at the very beginning (it is
actually the old `app_node`). This simulates the DFINITY-owned node of
the subnet. Note: it can never be behind one DKG interval because we are
never breaking it.
- The `upload_node` is the same as `download_state_node`, except if we
are doing recovery on failover nodes, in which case would simulate a
DFINITY-owned failover node.
- The `download_pool_node` is selected as such:
- In the happy path where a readonly SSH key can be deployed, it will be
the node with highest certification height. In that case, admin nodes
are just the `upload_node` (unchanged).
- Otherwise, the recovery coordinator will have to download the
consensus pool on the DFINITY-owned node (the only machine it can
access) with admin access, i.e. the `download_state_node`. It will be
able to replay up until the maximum certification height that it has on
that node, because it won't have the readonly key to download other
nodes' consensus pools. As [noted by
Leo](https://github.com/dfinity/ic/pull/7009#discussion_r2401712163),
this should not be a problem in practice. Therefore, we need to add the
`download_state_node` to the admin nodes to cover the case where it's
different from the `upload_node`, i.e. in a failover nodes scenario
(unchanged).