fix: do not update subnet available memory during install_code (#7432)
This PR fixes the replica implementation to not update (a temporary copy
of) the subnet available memory during `install_code`.

The change in this PR ensures that a canister does not fail during
`install_code`
- if it (temporarily) exceeds the subnet memory capacity during
pre-upgrade;
- due to seemingly exceeding the subnet memory capacity in
init/post-upgrade because a memory usage decrease after instantiating
the new WASM module is not reflected in (the temporary copy of) the
subnet available memory.

See the regression tests (that have been checked to fail on master) for
concrete examples of the above scenarios.

The change in this PR is safe because (the actual) subnet available
memory is updated when finishing execution of `install_code` based on
the overall change of canister's memory usage.

An alternative to this PR would be to properly update (the temporary
copy of) the subnet available memory during `install_code`, but this PR
is significantly simpler.