refactor(cketh/ckerc20): multi RPC results reduction (#7152)
([XC-508](https://dfinity.atlassian.net/browse/XC-508)) In preparation
for using the new
[`evm_rpc_client`](https://github.com/dfinity/evm-rpc-canister/tree/main/evm_rpc_client)
from the [EVM RPC canister
repository](https://github.com/dfinity/evm-rpc-canister) (in #7112),
this PR refactors the reduction mechanism for inconsistent RPC results
in the ckETH minter.

It introduces a new `ReductionStrategy` trait and a
`ToReducedWithStrategy` trait implemented on
`evm_rpc_types::MultiRpcResult` to allow reducing inconsistent
`MultiRpcResult` with a given `ReductionStrategy` with a fluent API.
This will make it easier to inline the current `EthRpcClient::eth_get*`
methods and use the corresponding methods on the `evm_rpc_client`
directly.

[XC-508]:
https://dfinity.atlassian.net/browse/XC-508?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ