chore: refactor CyclesAccountManager::refund_for_response_transmission (#7430)
This PR refactors the function
`CyclesAccountManager::refund_for_response_transmission` to take
`ic_types::messages::Payload` as input instead of
`ic_types::messages::Response` containing the former as one of its
fields. The motivation for this refactoring is to facilitate computing
the refund for response transmission in tests without having to
construct the entire struct `ic_types::messages::Response` of which only
a single field is relevant for the computation of the refund.