feat: use single bazel repository for mainnet canisters (#7421)
This changes the implementation of the `canisters` repository rule to
generate a single repository for all mainnet canisters.

This means the canisters are directly downloaded by the repo rule and we
don't need to create a separate `defs.bzl` that has to be imported.

This means that, when it is migrated to `MODULE.bazel`, the repo rule
can be re-used as is using `use_repo_rule`.

The first commit changes the implementation. The second commit is the
result of search and replace.

This appears to show that lots of canisters names were actually pointing
to the same canister. If this seems like a bug in the search and
replace, do let me know.