fix: startup time reporting: added hours, split line correctly (#7443)
Before, we didn't split the lines correctly and then the time was wrong
and the unit name included parts of the time (e.g.,
`11_663s_setup_permissions_service_`).
The fix adds support to also parse hours (this should not occur in
reality) and it splits the lines correctly by only taking the last part
as the name of the unit and using all previous parts as the time string.