fix: Make `two_subnets_simple` generate Application Subnets rather than System Subnets. (#7469)
`two_subnets_simple` and also `two_subnets_with_config` use
[SubnetType::Application](https://sourcegraph.com/github.com/dfinity/ic@8aba37dfe8c59f2a70ec2cce58a0dcf3eea98ec3/-/blob/rs/state_machine_tests/src/lib.rs?L4862)
to generate a `SubnetConfig` which would seem to imply that application
subnets are generated. However the subnet type is actually not contained
in the `SubnetConfig` rather this just initializes various limits to
those of an application subnet.
To make the generated subnets application subnets, the type must be
specified explicitly in the state machine builder, where the default is
[SubnetType::System](https://sourcegraph.com/github.com/dfinity/ic@8aba37dfe8c59f2a70ec2cce58a0dcf3eea98ec3/-/blob/rs/state_machine_tests/src/lib.rs?L1026).