fix(registry): Use optional flags init playload only in canisters (#7452)
Even though the init payload isn't used in mainnet (since we don't
expect init to be run again), it's used by many tools and libraries for
initializing a registry in local/testing environments. Therefore
introducing a breaking change (adding required fields) would break a lot
of use cases. However, since we have tests using the mainnet version of
the registry, it's difficult to change those fields into optional, since
the code change simultaneously change the encoder to send optional
flags, which the decoder - the mainnet version of registry cannot
decode. It's also risky to disable all those tests in order for the
change to make through. Therefore, in this PR we only change the decoder
(the registry canister) without changing the encoders. After this change
makes to the mainnet, we can clean up and change all the encoders to
also use the optional flags.

Note: candid can decode `T` as `opt T`, which is an important reason why
this approach works