feat: [MR-733] Generate `Refund` when dropping a message (#7440)
Generate a `Refund` whenever a best-effort canister message is timed out
or shed without a reject response having been directly enqueued into an
input queue. These `Refunds` are accumulated in the subnet's outbound
`RefundPool` but, for now, they are merely consumed at the end of the
round and the "cycles lost due to dropped messages" counter is increased
accordingly.