feat(registry): Revert restriction that unassigned nodes must have empty ssh_node_state_write_access. (#7473)
Reverts dfinity/ic#7314

When we made that change ☝️, we believed that there would be no
legitimate reason for an unassigned node to have a nonempty
`ssh_node_state_write_access`; we thought that such a configuration
could not be used in a constructive way. It seems that we did not ask
around whether there would ever a need for that.

After that PR ☝️, we realized that we should ask around, and when we
did, Consensus team told us that such an ability would be useful.
Therefore, this PR gets rid of enforcement of "if a node is unassigned,
`ssh_node_state_write_access` MUST be empty".

Closes [NNS1-4265].

[NNS1-4265]: https://dfinity.atlassian.net/browse/NNS1-4265