fix(tests): query every node to assert a subnet is healthy (#7422)
This PR queries every node after storing a new message in the message
canister in `assert_subnet_is_healthy`. This ensures that every node is
up-to-date and has the latest message when returning.
Otherwise, this is a source of flakiness where after quickly
[setting](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/nested/nns_recovery/common.rs?L227-234)
a new message with this function,
[breaking](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/nested/nns_recovery/common.rs?L283)
the subnet, and
[reading](https://sourcegraph.com/github.com/dfinity/ic/-/blob/rs/tests/nested/nns_recovery/common.rs?L284-290)
the message again, the old message would be returned because the node
that we are querying in this last read probably did not have the latest
state (including the new message).