chore: Exclude generated files when listing targets for CI
I believe it makes little sense to include them if their rule is already included, as this causes a problem: the `except attr(tags, ...)` exclusion does not apply to generated files (only their generating rules). Therefore, the generated file will be included in the output and built by CI even if its generating rule is marked with one of the excluded tags.