fix: Fix the release script for custom upgrades (#7319)
1. Supprot DRY_RUN
2. Fix the problem where clap thinks the entire
`--sns-governance-canister-id xxx-xxx` as the argument name

More about (2):

When there is an element "--sns-governance-canister-id xxx-xxx" in the
cmd (bash array), when it is evaluated as "${cmd[@]}" (with double
quotes), bash wouldn't do "word splitting" after expanding the array
(${cmd[@]}), and therefore "--sns-governance-canister-id xxx-xxx" is
treated as one single argument to pass into the ic-admin binary (and
then parsed by clap). In contrast, if the cmd is evaluated as
`${cmd[@]}` (no double quotes), then it doesn't matter if it has
`--sns-governance-canister-id xxx-xxx` or
`--sns-governance-canister-id=xxx-xxx`, since in the former case the
`--sns-governance-canister-id` and `xxx-xxx` are passed as 2 arguments.
We could potentially remove the double quotes, but using `--a=b` seems
better anyway.