test(outcalls): test socks proxy fallback  (#7371)
# Overview

This change enables the `canister_http_socks_test` by making it 
- Work
- Non flaky
- Run on CI (as a long test)

# Background
The test is meant to cover the
“[fallback](https://github.com/dfinity/ic/blob/0ad091870f3dc704a0a2a4a2063c54967d6e3d90/rs/https_outcalls/adapter/src/rpc_server.rs#L357)”
codepath of the http outcalls adapter. The adapter first tries to
connect directly to the target URL, and if that fails, it will try to do
so via one of the API boundary nodes (which has a running socks proxy
server to forward requests). This behaviour comes in very useful when
destinations are IPv4 only (as most nodes are IPv6 only, while API
boundary nodes have both IPv6 and IPv4 connectivity and can successfully
bridge the gap as proxies).

All http outcalls test spawn a UVM and on it, spawn a minimal httpbin
server which can handle requests. Then the outcalls are made to this
server.

The test was originally designed when outcalls used old boundary nodes,
(not the API boundary nodes) as a proxy, and they had different
architecture (concretely they could easily support IPv4 connectivity)

The test has been disabled for a while because there were a few issues
with (in short, it wasn’t working):
1. Nodes running on farm (including API boundary nodes) can’t easily
support IPv4 connectivity. Meaning that a simple (ipv6)node <->
(ipv6/ipv4)api_bn <-> (ipv4)http_bin can’t work
2. Because of some subtle bugs in the client libraries which create
socks requests (reqwest, hyper-util, hyper_socks2 etc), requests are
always bundled as “domain names” (with
[atyp](https://datatracker.ietf.org/doc/html/rfc1928) = 0x03), even if
they actually are IPv6 IPs. This would make the server try to dns
resolve those IPv6 URLs, and will fail to parse them. Note that in
production this would only affect requests that are IPv6 only, and fail
on the direct path. However such requests would anyway probably fail on
the proxy path too (as the proxy path is designed for IPv4 requests).
3. There are [firewall rules set
up](https://github.com/dfinity/ic/blob/c46f1b371f0d67a19ba5af5fd6b890f8c81771d1/rs/ic_os/config/templates/ic.json5.template#L290)
that block outbound connections coming from the outcalls adapter to a
range of IP addresses including production nodes (but not production API
boundary nodes). This is to prevent canisters using the adapter to reach
endpoints that should only be accessible to replicas. This makes it so
even system test api boundary nodes that fall in that IP range will be
blocked and hence can’t be used as proxies. Actually, even the UVM
itself can fall under this IP range, however, as only ports 1-19,999
were blocked, an easy “hack” was to install the httpbin server on [port
20443](https://github.com/dfinity/ic/blob/c46f1b371f0d67a19ba5af5fd6b890f8c81771d1/rs/tests/networking/canister_http/canister_http.rs#L234).
4. API boundary nodes also have firewall rules (and, in addition, Dante
config rules) which disallow it to connect to almost all ports, except
for 443 and 80 (generally belonging to https and http). This makes the
hack of problem 3 no longer effective.

# Solution
To overcome those problems, we are doing the following:
- Make httpbin run on 443 (https) instead of 20443. This directly fixes
problem 4, but opens problem 3 again
- When installing httpbin, we ssh into all (subnet) nodes and insert an
nft accept rule to the uvm’s 443 port, which overwrites the firewall
rules described with problem 3.
- Similarly, we ssh into all nodes and insert similar nft accept rules
but this time to API boundary nodes. This, together with the previous
point completely solve problem 3.
- The canister makes an outcall via a [nip.io](http://nip.io/)
subdomain. (see more in [nip.io](http://nip.io/)). Basically a domain
name like `2a00-1450-4001-80e--200e.ipv6.nip.io` is dns resolved to
`[2a00:1450:4001:80e::200e]`. The socks server will successfully dns
resolve the domain name to the right IP address. In addition, we had to
make sure to also serve TLS certificates for that domain name from the
http bin via minica. This solves problem 2.
- Artificially block the node <-> uvm connection with an nft reject rule
towards the uvm’s port 443. This makes the direct request fail
immediately, and fall back to the proxied connection which succeeds.
This fixes problem 1

The canister calls out to an `/ip` endpoint on the httpbin, which
returns the IP address of the caller. That IP is either that of the node
(in the case of direct connections) or of the API boundary node’s (in
the case of proxied connections).

# Conclusion
These changes make the system test finally work as intended. This test
is needed in preparation for changes to the app subnets being able to
also use api boundary nodes as proxies.

In the long run, we plan to fix those issues in a less "hacky way":
- fix client libraries upstream
- develop infrastructure to be able to start IC nodes with IPv4
connectivity
- have a test-only firewall config
- use a custom dns patch instead of relying on nip.io

However these have a significant overhead, and we'd like to be able to
have the test ready for migrating app subnets.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>