chore: migration canister passes the caller to mgmt canister (#7496)
This PR makes the migration canister pass the caller of a renaming
request to the management canister executing that request.

This PR is the first step towards storing the actual caller of a
renaming request in canister history that currently only stores the
migration canister as the "caller" since canister history is updated by
the management canister and the management canister only sees the
migration canister as the caller (effectively hiding the actual caller
who initiated canister renaming).

Before making the management canister require the actual caller in its
input payload to the method `rename_canister`, we have to update its
clients first and the only allowed client of this (internal) management
canister API is the migration canister.

The full roadmap is as follows:
- migration canister sends the actual caller in its payload to the
method `rename_canister` of the management canister and the management
canister ignores that extra field for now - this PR;
- the migration canister change is deployed (the canister is
upgraded/reinstalled) on the ICP mainnet with an allowlist (no public
release of the canister ID migration feature yet) - a follow-up NNS
proposal;
- management canister requires the actual caller in its input payload to
the method `rename_canister` and stores that caller in the canister
history - a follow-up PR;
- the management canister change is rolled-out on the ICP mainnet - a
regular subnet upgrade release;
- the allowlist in the migration canister is removed (this is the public
release of the canister ID migration feature) - a follow-up NNS
proposal.