fix: Address "VM is running" check issues (#7458)
### is_running
`wait_for_shutdown` used to call `is_running` which returned true iff
the domain existed and was in `VIR_DOMAIN_RUNNING` state. However, we've
observed that the code intermittently misbehaves.

According to our theory:
  - the VM is in `VIR_DOMAIN_SHUTDOWN` state
  - `is_running` return false
  - `wait_for_shutdown` returns
  - we try to clean up the temporary files which fails

The new logic:
  - If domain exists and running: all good (no changes here)
  - If domain exists but in crashed state: consider dead and log
  - If domain exists and is in any other state: consider running and log
- If domain does not exist or its state cannot be queried: consider dead
and log

### try_destroy_existing_vm
We change the implementation to include a call to undefine. This may
help remove the VM if it's in an odd state where it's defined but is not
running.