chore: move mainnet images to MODULE.bazel (#7531)
This is required in order to get rid of WORKSPACE.bazel.

This also fixes the mainnet icos versions repo rule (ensures the json
file is watched) and extracts the `MAINNET_` variables out of the
mainnet-icos-images (avoids the dependency on mainnets icos versions
repo rule which would create a dependency cycle issue).