test: Use mainnet version of migration canister in tests (#7497)
This commit updates the the test framework, pocketIC, and release
scripts to use the mainnet version of the migration canister.
Previously, the mainnet version of the migration canister was not
compiled from any commit of the repo, such that it was not possible yet
to fully integrate it into all of the test framework and the release
script. This is no longer the case, and this commit can now treat the
canister like any other NNS canister by referring to its mainnet git
commit.

Furthermore, this commit adds the migration canister to the release
script. This streamlines future deployments and e.g. makes it possible
to update the mainnet commit id in an automated way.