fix(IDX): handle merge queue checks differently (#7536)
The crux of the issue is this: by default status checks are run on both
PRs and merge queues. However, the `ci-pr-only` workflow only runs on
PRs. So if no status check is posted for this workflow the merge queue
will be stuck.

This PR introduces a CI job to automatically post the commit status for
`ci-pr-only` as successful whenever a merge queue is triggered. It's a
little hacky, but I think a little clearer than the if / else logic I
had before.