test(node): add nested GuestOS upgrade orchestrator check and diagnostics (#7498)
NODE-1761

The nested GuestOS upgrade tests only check for the final GuestOS
version—not the Orchestrator health. This is why we haven’t seen
flakiness on these tests, but have on the HostOS upgrade tests. This
adds the orchestrator check to the nested GuestOS upgrade tests along
with diagnostic logging in failures.