docs(ck_testnet): FI-1907: Bump cycles_for_archive_creation from 1T to 10T (#7532)
Bump the value of `cycles_for_archive_creation` from 1T to 10T in the ck
token testnet docs. For ledgers running on larger subnets, more than 1T
cycles may be required for spawning an archive (e.g., on a 34 node
subnet, 1.3T is currently needed for `create_canister`, on top of which
more cycles are needed to install the wasm, process blocks, and for the
freezing threshold).