chore: A more secure and reliable way of checking SEV availability (#7534)
Previously, we checked the existence of /dev/sev-guest and made cpuid
calls to check whether SEV is enabled. This was not perfect because
/dev/sev-guest may not be available very early and cpuid calls can be
intercepted by a malicious host. This new approach stores the result in
an environment variable called `SEV_ACTIVE`. The benefit is that 1) it's
available when systemd starts 2) only queries cpuid once and stores the
result so every service sees a consistent value.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Andrew Battat <113942931+andrewbattat@users.noreply.github.com>