fix: data races during canister migration validation (#7549)
This PR fixes data races during canister migration validation.

So far the validation logic checked if the source and target canisters
are not involved in a migration by checking existing requests,
performing asynchronous calls, and finally creating a new request.
However, this pattern is prone to concurrency bugs if multiple requests
to migrate the source or target canisters are submitted concurrently.

This PR fixes the data race by locking both the source and target before
checking existing requests and performing asynchronous calls. The locks
are only acquired after checking if the caller is authorized to prevent
unauthorized callers from acquiring the locks and blocking canister
migration for authorized callers.

Additionally, this PR performs the following changes:
- `canister_status` calls are performed with the management canister
being the callee (this is fine since such calls are not performed during
canister migration).