feat: Monitor from the HostOS whether GuestOS boot was successful using the serial output (#7507)
We use the serial output as the means of communication because it's very
reliable - we get information even if the GuestOS boot fails early.

The new log-boot-{failure,success} services print a message on the
serial console which gets picked up by the GuestVM runner in the HostOS.
If a failure happens or no success is reported within the timeout, the
HostOS prints the GuestOS serial output.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>
Co-authored-by: Andrew Battat <113942931+andrewbattat@users.noreply.github.com>