chore: Move NNS public key under /run and refactor config bootstraping (#7388)
This allows us to access the NNS public key (from rootfs or injected)
even when the encrypted data partition is not available. To allow
populating the NNS key without depending on the encrypted data
partition, it had to be moved out the bootstrap-ic-node.service.

The logic was moved to a new command (`config populate-nns-key`) which
is invoked in init-config.service.

To avoid having to untar the config tar twice, the untarring is now done
in init-config.service (which now also verifies that the tar only
contains regular files and directories *before* untarring) and the
untarred contents are stored under /run/config/bootstrap.

---------

Co-authored-by: Andrew Battat <113942931+andrewbattat@users.noreply.github.com>
Co-authored-by: Andrew Battat <andrew.battat@dfinity.org>