fix(governance): Release the lock for minting node provider rewards... (#7530)
... even in sad cases, such as when the Node Rewards canister is down
(e.g. for upgrades). This is done by following Scott Mayer's excellent
advice from [Effective C++][eff]:

> Item 13: Use objects to manage resources.

[eff]:
https://ptgmedia.pearsoncmg.com/images/9780321334879/samplepages/0321334876.pdf

# Background

This subtle bug was brought to light by a [failure] of
`upgrade_canisters_with_golden_nns_state.rs` (testing FTW!). The smoking
gun in the logs was

```
2025-12-04 01:51:32.782197895 UTC: [Canister rrkah-fqaaa-aaaaa-aaaaq-cai] [Governance] Error when minting monthly node provider rewards in run_periodic_tasks: External: Error calling 'get_node_providers_monthly_xdr_rewards': code: Some(5), message: IC0509: Canister sgymv-uiaaa-aaaaa-aaaia-cai is not running
```

Thanks to Jason for locating this needle in the haystack! (With the
benefit of hindsight, this could have "easily" been located by searching
backwards from the end of the logs for "error".)

[failure]: https://github.com/dfinity/ic/actions/runs/19020197264

# Secondary Change(s)

Changed description of `ErrorType::Unavailable` to reflect how it is
usually used in practice: when another instance of the operation is
already in flight.

# Testing

If you check out
[inject-fault-into-get_monthly_node_provider_rewards-daniel-wong][inject-fault]
(one commit away from `master`, and then run
`upgrade_canisters_with_golden_nns_state`, it fails. But if you cherry
pick the one commit in
`inject-fault-into-get_monthly_node_provider_rewards-daniel-wong` from
this branch, then, the test passes.

[inject-fault]: https://github.com/dfinity/ic/pull/7552

# Future Work

We should isolate different pieces of background work from one another.
Ideally, each kind of background work would be done using a separate
timer, rather trying to schedule them in a monolithic heartbeat
function. I think we resolved to do this a while ago, but I guess we
never found the time.