feat: [NODE-1727] Do not cleanup tmpdir if umount fails (#7492)
When tempdir is dropped, it calls `remove_dir_all`. We already try to
drop and unmount the partition before tempdir is cleaned up, but if the
unmount fails, the tempdir cleanup removes all contents.

Now, we retry the unmount a few times if the resource is busy, and
ultimately if the drop fails, we do not try to cleanup the tmpdir.

---------

Co-authored-by: David Frank <david.frank@dfinity.org>