feat(outcalls): add price_version parameter to the http canister request args (#7447)
This PR adds the price version, as described in
https://github.com/dfinity/portal/pull/6101 to the management canister
`http_request` endpoint.

This PR also persists the version in the replicated state. In the future
that will be used to determine which pricing strategy should be applied.

The flag is a candid opt nat32, which, if missing, defaults to "1",
which corresponds to the "current" pricing. Similarly, if set to "2",
the flag will default back to 1, as version 2 is not implemented yet.

In the future, we will support pricing_Version = 2, which corresponds to
the "pay-as-you-go" pricing.

---------

Co-authored-by: kpop-dfinity <125868903+kpop-dfinity@users.noreply.github.com>