feat(node-reward-canister): Allow clients to backtrack rewards  (#7591)
This pull request updates the `get_node_providers_rewards_calculation`
API types to use f64 instead of u64 for reward results.
The change allows clients to independently reproduce and verify the
reward calculations, which isn’t possible when using only integer (u64)
values.

For the same reason, we also aim to ensure that daily rewards are fully
independent—so that the total amount minted by governance for a reward
period exactly matches the sum of its individual daily rewards.
To achieve this, the final aggregated `DailyNodeProviderRewards`
`rewards_total_xdr_permyriad` is represented as a u64.

---------

Co-authored-by: IDX GitHub Automation <infra+github-automation@dfinity.org>