feat(nervous_system): Add `canister_metadata` to `ManagementCanisterClient` (#7566)
Add `canister_metadata` to `ManagementCanisterClient` so that it can be
called by the governance canister. They are not used yet.

In the future, NNS Governance will use this endpoint to fetch .did files
from the target canisters of ExecuteNnsFunction and use the .did files
to translate payload blobs to meaningful data.

The `canister_metadata` management canister call is equivalent to `dfx
canister metadata $CANISTER $METADATA_NAME`