test(nns-recovery): add sequential NP actions system test (#7543)
This PR adds a new NNS recovery system test where the actions of Node
Providers are simulated in sequence, instead of in parallel.

Notable change: `simulate_node_provider_action` now also checks that the
node reports the new version. In the sequential case, this ensures that
we fix a new node only when the previous one has already rebooted on the
correct version. In the parallel case, this does not really change
anything, since we anyways also do the check afterwards in
`assert_subnet_is_healthy`.

This PR also removes the overall timeout on NNS recovery system tests
after realizing that the `setup` task does not actually take that much
time to complete.