chore: refactoring test to allow turning on the node swaps feature (#7616)
Initially, this test was there to protect from accidental enabling of
node swaps on the network.

Now that we are preparing for launching the feature, this PR refactors
the test to deploy a `test` configuration of the canister, which
contains the code to override the flag for swaps being enabled. By
default, swaps are going to be disabled in tests.