fix(ckbtc): time estimate for finalization (#7621)
The ckBTC minter estimates how long it takes for a withdrawal
transaction to receive the sufficient number of confirmations to be
considered finalized by multiplying the average block time with the
required number of confirmations. For unknown reasons, the block time
for Bitcoin testnet was set to 1 minute instead of 10 minutes (same
block time as Bitcoin mainnet).