chore: Clean up DID checks (#7588)
This simplifies the candid compatibility checks by using a macro (with
`sh_test`) instead of a full blown rule.

This also fixes an issue with the "patched" test candids which performed
checks against non-existent files.

This also moves the checks to a new job that only runs on PRs. The
checks are not meaningful without a merge base. The checks are still run
on non-PRs but only against HEAD, effectively making sure the test
infrastructure works.

---------

Co-authored-by: mraszyk <31483726+mraszyk@users.noreply.github.com>