fix: no data certificate is present in canister http transform (#7562)
This PR makes the data certificate (a threshold-signed IC certificate
for the value of a canister's certified variables) not available in the
canister http transform function.

This means that the system API `ic0.data_certificate_present` returns
`0` and calling the system APIs `ic0.data_certificate_size` and
`ic0.data_certificate_copy` traps. The system API
`ic0.in_replicated_execution` still returns `false` for canister http
transform.

The justification for this change is that the (transformed) canister
http response is consumed by the canister itself which does not need an
IC certificate for the value of its own certified variables (unlike an
external user).

The motivation for this change is that the legacy format of the routing
table present in the IC certificate delegation will be deprecated in the
future and thus a breaking change to the format of the IC certificate is
inevitable. Since there's no fundamental motivation for using the IC
certificate in canister http transform functions, this PR makes the IC
certificate unavailable in canister http transform functions instead of
implementing a migration path towards the new format of IC certificates
in canister http transform functions.