test: Stop NNS delegations test breaking with new System APIs (#7623)
New system APIs are normally added to the universal canister. This broke
the NNS delegations test. The test tried to install the universal
canister on the mainnet replica version, which breaks since the
universal canister imports the new System APIs which do not yet exist on
the mainnet replica.

Switch the test to use a handrolled canister for checking delegations.