refactor(IDX): remove anchor setup step (#7512)
Looking at the UI, it does bother me a little that we have two jobs
`Anchor Setup` that are never run. This refactors the github actions CI
code to use anchors the way GitHub intends them to be used - defining
them in the first job and re-using them in subsequent jobs.