chore: Extend log message for invalid stream slices during block making (#7658)
The main reason for invalid slices (from honest subnets) is attempting
to build a block relative to a validation context referencing a
certified state that is too far behind the most recent state. Having
information in the logs on how far behind the referenced certified
height is should make debugging easier.