fix(ckdoge): finalization of transactions sent by the minter (#7589)
Follow-up on #7360 to ensure that transactions on the Dogecoin
blockchain sent by the minter in case of a withdrawal:

1. Will be considered finalized by the minter once a sufficient number
of confirmations is reached.
2. Can be resubmitted if the sent transaction was not confirmed within a
certain time window.
3. Can be cancelled in case the transaction would use too many inputs
and be at risk of being non-standard.

This PR also improves the test infrastructure from #7499 and #7577 to do
some batching operations (like pushing UTXOs to the mock
Bitcoin/Dogecoin canister with a single call) to reduce the runtime of
integration tests that need a large number of UTXOs.

`CI_OVERRIDE_DIDC_CHECK`: The only change in the canister Candid API is
for `ic-bitcoin-canister-mock`, which is a canister solely used for
testing purposes.