chore(crypto): CRP-2456 Remove ed25519_public_key_from_der
This function is quite error prone since it doesn't check anything
about its argument; not even checking that the removed bytes match the
expected DER header, much less that the returned value is actually a
valid public key.

Happily this function is not used anymore within the monorepo. Given
this, just remove it, so that nobody uses it in the future.