chore: Fix flaky VetKD max payload size test (#7665)
In this test, there are enough VetKD shares for the payload builder to
aggregate them into two agreements. At the same time, the test sets the
max payload size low enough, such that only one of the agreements fits
into the payload.

Before https://github.com/dfinity/ic/pull/7373 introduced
multi-threading to the VetKD payload builder, the agreement that ended
up in the payload was always the one for the first request (as they were
handled sequentially). The test asserted that this was indeed the case.

However, after introducing multi-threading it can happen that only the
agreement for the second request is included into the payload, instead.
This PR fixes the resulting flakiness by accepting either agreement, as
long as there is only one.